links on stubs, busines logic here we go!

+++ src/ui/CodeBlockManager.cpp
@@ -1,10 +1,9 @@
-#include "WindowManager.hpp"
 #include "bit_set/bit_set.hpp"
-#include "utility.hpp"
-#include "TextWindow.hpp"
+#include "ui/code-blocks/CodeBlock.hpp"
+#include "ui/code-blocks/CodeBlockManager.hpp"
+#include "utils/utility.hpp"
 #include <SFML/Graphics/Rect.hpp>
 #include <SFML/System/Vector2.hpp>
-// #include "../include/multimax.hpp"
 #include <algorithm>
 #include <cmath>
 #include <numeric>
@@ -21,7 +20,10 @@ class Transformable;
 namespace Canvas {
 auto CodeBlocksManager::get_all_windows() -> BlockSet {
   // since findWindows already checks for nullopt we just want any window
-  return findBlocks([](auto const &window) {(void)window; return true; });
+  return findBlocks([](auto const &window) {
+    (void)window;
+    return true;
+  });
 }
 
 auto CodeBlocksManager::newCodeBlock(sf::String &&header, sf::String &&source)
@@ -34,7 +36,7 @@ auto CodeBlocksManager::newCodeBlock(sf::String &&header, sf::String &&source)
   // will filter them with this
   // we also make sure we don't include ourselves in the intersection group
   BlockSet available_windows = get_all_windows();
-  available_windows.erase(uint(window_it - m_windows.begin()));
+  available_windows.erase(int(window_it - m_windows.begin()));
 
   auto where_to_place =
       findOpenspace({target_size.width, target_size.height}, available_windows);
@@ -50,7 +52,7 @@ auto CodeBlocksManager::deleteCodeBlock(BlockSet windows) -> void {
 }
 
 auto CodeBlocksManager::draw(sf::RenderTarget &target,
-                         sf::RenderStates states) const -> void {
+                             sf::RenderStates states) const -> void {
   for_each_optional(m_windows, [&target, &states](auto const &window) {
     states.transform = window->getTransform();
     target.draw(*window, states);
@@ -58,7 +60,8 @@ auto CodeBlocksManager::draw(sf::RenderTarget &target,
 }
 
 auto CodeBlocksManager::findOpenspace(sf::Vector2f rec_size,
-                                  BlockSet available_windows) -> sf::Vector2f {
+                                      BlockSet available_windows)
+    -> sf::Vector2f {
   // no windows? no problem!
   if (available_windows.empty())
     return {0, 0};
@@ -67,7 +70,7 @@ auto CodeBlocksManager::findOpenspace(sf::Vector2f rec_size,
   float min_x = INFINITY;
   float min_y = INFINITY;
   for (auto window : available_windows) {
-    auto gb = m_windows[window]->getGlobalBounds();
+    auto gb = m_windows[uint(window)]->getGlobalBounds();
     min_x = std::min(min_x, gb.left);
     min_y = std::min(min_y, gb.top);
   }
@@ -93,7 +96,7 @@ auto CodeBlocksManager::findOpenspace(sf::Vector2f rec_size,
     float max_x = -INFINITY;
     float max_y = -INFINITY;
     for (auto intersecter : intersects) {
-      auto bounds = m_windows[intersecter]->getGlobalBounds();
+      auto bounds = m_windows[uint(intersecter)]->getGlobalBounds();
       max_x = std::max(max_x, bounds.left + bounds.width);
       max_y = std::max(max_y, bounds.top + bounds.height);
     }
